% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposal.R
\name{accept_workflow_proposal}
\alias{accept_workflow_proposal}
\title{Accept all or selected nodes from a workflow proposal}
\usage{
accept_workflow_proposal(
  proposal,
  graph = list(nodes = list(), edges = list()),
  node_ids = NULL
)
}
\arguments{
\item{proposal}{Workflow proposal.}

\item{graph}{Existing graph to merge into.}

\item{node_ids}{Optional proposed node identifiers to accept.}
}
\description{
Accept all or selected nodes from a workflow proposal
}
