% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser_controls.R
\name{browser_alert}
\alias{browser_alert}
\title{Browser-native inline alert}
\usage{
browser_alert(
  title,
  message = NULL,
  status = c("info", "success", "warning", "error")
)
}
\arguments{
\item{title}{Alert title.}

\item{message}{Alert message.}

\item{status}{Alert tone.}
}
\description{
Browser-native inline alert
}
