% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser_surfaces.R
\name{browser_breadcrumbs}
\alias{browser_breadcrumbs}
\title{Browser-native breadcrumbs}
\usage{
browser_breadcrumbs(id, items)
}
\arguments{
\item{id}{Stable component identifier.}

\item{items}{List of id and label records.}
}
\description{
Browser-native breadcrumbs
}
