% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser_controls.R
\name{browser_skeleton}
\alias{browser_skeleton}
\title{Browser-native loading skeleton}
\usage{
browser_skeleton(lines = 3L, label = "Loading")
}
\arguments{
\item{lines}{Number of placeholder lines.}

\item{label}{Accessible loading label.}
}
\description{
Browser-native loading skeleton
}
