% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser_surfaces.R
\name{browser_tooltip}
\alias{browser_tooltip}
\title{Browser-native tooltip}
\usage{
browser_tooltip(id, trigger, content, placement = "top")
}
\arguments{
\item{id}{Stable component identifier.}

\item{trigger}{Trigger content.}

\item{content}{Tooltip content.}

\item{placement}{Preferred placement.}
}
\description{
Browser-native tooltip
}
