% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime.R
\name{cancel_workflow_branch}
\alias{cancel_workflow_branch}
\title{Cancel a node and its dependent branch}
\usage{
cancel_workflow_branch(runtime, node_id)
}
\arguments{
\item{runtime}{Runtime returned by `workflow_runtime()`.}

\item{node_id}{Branch root.}
}
\description{
Cancel a node and its dependent branch
}
