% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget.R
\name{capability_canvas_output}
\alias{capability_canvas_output}
\title{Shiny output binding for capability_canvas}
\usage{
capability_canvas_output(output_id, width = "100\%", height = "640px")
}
\arguments{
\item{output_id}{Shiny output identifier.}

\item{width, height}{Widget dimensions.}
}
\description{
Shiny output binding for capability_canvas
}
