% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{capability_registry_add}
\alias{capability_registry_add}
\title{Add a capability to a registry}
\usage{
capability_registry_add(registry, capability)
}
\arguments{
\item{registry}{Capability registry.}

\item{capability}{Capability definition.}
}
\description{
Add a capability to a registry
}
