% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{capability_registry_get}
\alias{capability_registry_get}
\title{Retrieve a capability}
\usage{
capability_registry_get(registry, id)
}
\arguments{
\item{registry}{Capability registry.}

\item{id}{Capability identifier.}
}
\description{
Retrieve a capability
}
