% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{capability_registry_list}
\alias{capability_registry_list}
\title{List capabilities deterministically}
\usage{
capability_registry_list(registry)
}
\arguments{
\item{registry}{Capability registry.}
}
\description{
List capabilities deterministically
}
