% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_component_transport.R
\name{direct_component_output}
\alias{direct_component_output}
\title{Direct component Shiny output}
\usage{
direct_component_output(
  output_id,
  component,
  width = "100\%",
  height = "420px"
)
}
\arguments{
\item{output_id}{Shiny output id.}

\item{component}{Registered direct component name.}

\item{width, height}{CSS dimensions.}
}
\description{
Direct component Shiny output
}
