% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{expand_workflow_composite}
\alias{expand_workflow_composite}
\title{Expand one workflow composite}
\usage{
expand_workflow_composite(graph, composite_id)
}
\arguments{
\item{graph}{Workflow graph.}

\item{composite_id}{Composite identifier.}
}
\description{
Expand one workflow composite
}
