% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_fingerprint}
\alias{graph_fingerprint}
\title{Fingerprint a normalized graph}
\usage{
graph_fingerprint(graph)
}
\arguments{
\item{graph}{Workflow graph list.}
}
\description{
Fingerprint a normalized graph
}
