% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser_surfaces.R
\name{notification_center}
\alias{notification_center}
\title{Browser-native notification center}
\usage{
notification_center(
  id,
  notifications = list(),
  max_visible = 4L,
  max_history = 100L
)
}
\arguments{
\item{id}{Stable component identifier.}

\item{notifications}{List of notification records.}

\item{max_visible}{Maximum visible toast count.}

\item{max_history}{Maximum retained history count.}
}
\description{
Browser-native notification center
}
