% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget.R
\name{render_capability_canvas}
\alias{render_capability_canvas}
\title{Shiny renderer for capability_canvas}
\usage{
render_capability_canvas(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{Expression returning a widget.}

\item{env}{Evaluation environment.}

\item{quoted}{Whether \code{expr} is quoted.}
}
\description{
Shiny renderer for capability_canvas
}
