% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_component_transport.R
\name{render_direct_component}
\alias{render_direct_component}
\title{Render a direct component}
\usage{
render_direct_component(
  expr,
  output_func,
  env = parent.frame(),
  quoted = FALSE
)
}
\arguments{
\item{expr}{Expression returning a `shinycapabilities_direct_component`.}

\item{output_func}{Output constructor.}

\item{env}{Evaluation environment.}

\item{quoted}{Whether `expr` is quoted.}
}
\description{
Render a direct component
}
