% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{restore_workflow_document}
\alias{restore_workflow_document}
\title{Restore a workflow document}
\usage{
restore_workflow_document(value)
}
\arguments{
\item{value}{JSON text or decoded list.}
}
\description{
Restore a workflow document
}
