% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_activity_monitor.R
\name{run_agent_activity_monitor_demo}
\alias{run_agent_activity_monitor_demo}
\title{Run the Agent Activity Monitor demo}
\usage{
run_agent_activity_monitor_demo(...)
}
\arguments{
\item{...}{Arguments passed to [shiny::runApp()].}
}
\description{
Run the Agent Activity Monitor demo
}
