% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser_controls.R
\name{run_browser_controls_gallery}
\alias{run_browser_controls_gallery}
\title{Run the browser controls gallery}
\usage{
run_browser_controls_gallery(...)
}
\arguments{
\item{...}{Arguments passed to [shiny::runApp()].}
}
\description{
Run the browser controls gallery
}
