% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser_surfaces.R
\name{run_browser_surface_gallery}
\alias{run_browser_surface_gallery}
\title{Run the complete browser surface gallery}
\usage{
run_browser_surface_gallery(launch.browser = interactive())
}
\arguments{
\item{launch.browser}{Passed to [shiny::runApp()].}
}
\description{
Run the complete browser surface gallery
}
