% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{run_capability_demo}
\alias{run_capability_demo}
\title{Run the bundled demonstration Shiny app}
\usage{
run_capability_demo(...)
}
\arguments{
\item{...}{Arguments passed to \code{shiny::runApp()}.}
}
\description{
Run the bundled demonstration Shiny app
}
