% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_editor.R
\name{run_code_editor_demo}
\alias{run_code_editor_demo}
\title{Run the Monaco Editor 1.0 demo}
\usage{
run_code_editor_demo(...)
}
\arguments{
\item{...}{Arguments passed to [shiny::runApp()].}
}
\description{
Run the Monaco Editor 1.0 demo
}
