% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_grid.R
\name{run_data_grid_demo}
\alias{run_data_grid_demo}
\title{Run the analytical data-grid demo}
\usage{
run_data_grid_demo(...)
}
\arguments{
\item{...}{Arguments passed to [shiny::runApp()].}
}
\description{
Run the analytical data-grid demo
}
