% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command_palette_direct.R
\name{run_direct_transport_demo}
\alias{run_direct_transport_demo}
\title{Run the direct transport comparison demo}
\usage{
run_direct_transport_demo(...)
}
\arguments{
\item{...}{Arguments passed to [shiny::runApp()].}
}
\description{
Run the direct transport comparison demo
}
