% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execution_replay.R
\name{run_execution_replay_demo}
\alias{run_execution_replay_demo}
\title{Run the Execution Replay demo}
\usage{
run_execution_replay_demo(...)
}
\arguments{
\item{...}{Arguments passed to [shiny::runApp()].}
}
\description{
Run the Execution Replay demo
}
