% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_components.R
\name{run_interaction_components_demo}
\alias{run_interaction_components_demo}
\title{Run the interaction-components demo}
\usage{
run_interaction_components_demo(...)
}
\arguments{
\item{...}{Arguments passed to [shiny::runApp()].}
}
\description{
Run the interaction-components demo
}
