% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_inspector.R
\name{run_object_inspector_demo}
\alias{run_object_inspector_demo}
\title{Run the Structured Object Inspector 1.0 demo}
\usage{
run_object_inspector_demo(...)
}
\arguments{
\item{...}{Arguments passed to [shiny::runApp()].}
}
\description{
Run the Structured Object Inspector 1.0 demo
}
