% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_workbench.R
\name{run_parameter_workbench_demo}
\alias{run_parameter_workbench_demo}
\title{Run the Typed Parameter Workbench demo}
\usage{
run_parameter_workbench_demo(...)
}
\arguments{
\item{...}{Arguments passed to [shiny::runApp()].}
}
\description{
Run the Typed Parameter Workbench demo
}
