% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistent_ui.R
\name{run_persistent_ui_demo}
\alias{run_persistent_ui_demo}
\title{Run the Persistent Dynamic UI comparison demo}
\usage{
run_persistent_ui_demo(...)
}
\arguments{
\item{...}{Arguments passed to [shiny::runApp()].}
}
\description{
Run the Persistent Dynamic UI comparison demo
}
