% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationship_graph.R
\name{run_relationship_graph_demo}
\alias{run_relationship_graph_demo}
\title{Run the Relationship Graph demo}
\usage{
run_relationship_graph_demo(...)
}
\arguments{
\item{...}{Arguments passed to [shiny::runApp()].}
}
\description{
Run the Relationship Graph demo
}
