% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_pane.R
\name{run_split_pane_demo}
\alias{run_split_pane_demo}
\title{Run the accessible split-pane demo}
\usage{
run_split_pane_demo(...)
}
\arguments{
\item{...}{Arguments passed to [shiny::runApp()].}
}
\description{
Run the accessible split-pane demo
}
