% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{serialize_workflow_document}
\alias{serialize_workflow_document}
\title{Serialize a workflow document}
\usage{
serialize_workflow_document(document)
}
\arguments{
\item{document}{Workflow document.}
}
\description{
Serialize a workflow document
}
