% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{shinycapabilities_icon_allowlist}
\alias{shinycapabilities_icon_allowlist}
\title{Allowed neutral icon identifiers}
\usage{
shinycapabilities_icon_allowlist()
}
\value{
A character vector of locally rendered icon identifiers.
}
\description{
Allowed neutral icon identifiers
}
