% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser_surfaces.R
\name{update_browser_surface}
\alias{update_browser_surface}
\title{Update a browser surface}
\usage{
update_browser_surface(
  session = shiny::getDefaultReactiveDomain(),
  id,
  action,
  ...
)
}
\arguments{
\item{session}{Shiny session.}

\item{id}{Surface identifier.}

\item{action}{Bounded update action.}

\item{...}{Action payload.}
}
\description{
Update a browser surface
}
