% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser_surfaces.R
\name{update_notification_center}
\alias{update_notification_center}
\title{Update the notification center}
\usage{
update_notification_center(
  session = shiny::getDefaultReactiveDomain(),
  id,
  notifications = list(),
  mode = c("replace", "append", "dismiss")
)
}
\arguments{
\item{session}{Shiny session.}

\item{id}{Surface identifier.}

\item{notifications}{Notification records.}

\item{mode}{replace, append, or dismiss.}
}
\description{
Update the notification center
}
