% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_input.R
\name{update_selection_input}
\alias{update_selection_input}
\title{Update an Analytics Selection System input}
\usage{
update_selection_input(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  ...
)
}
\arguments{
\item{session}{Shiny session.}

\item{inputId}{Input identifier.}

\item{...}{Public model fields to update.}
}
\description{
Update an Analytics Selection System input
}
