% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_pane.R
\name{update_split_pane}
\alias{update_split_pane}
\title{Update an accessible split pane}
\usage{
update_split_pane(
  session = shiny::getDefaultReactiveDomain(),
  input_id,
  sizes = NULL,
  collapse = NULL,
  expand = NULL,
  reset = FALSE
)
}
\arguments{
\item{session}{Active Shiny session.}

\item{input_id}{Split-pane input identifier.}

\item{sizes}{Optional named or ordered percentage layout.}

\item{collapse, expand}{Pane ids to collapse or expand.}

\item{reset}{Restore initial sizes.}
}
\description{
Update an accessible split pane
}
