% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{validate_workflow}
\alias{validate_workflow}
\title{Validate a workflow graph}
\usage{
validate_workflow(registry, graph)
}
\arguments{
\item{registry}{Capability registry.}

\item{graph}{Workflow graph.}
}
\description{
Validate a workflow graph
}
