% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{workflow_document}
\alias{workflow_document}
\title{Create a serializable workflow document}
\usage{
workflow_document(graph, output_placements = list())
}
\arguments{
\item{graph}{Workflow graph.}

\item{output_placements}{Secondary output placements.}
}
\description{
Create a serializable workflow document
}
