% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{capability_canvas_server}
\alias{capability_canvas_server}
\title{Capability workflow module server}
\usage{
capability_canvas_server(
  id,
  registry,
  initial_graph = list(nodes = list(), edges = list()),
  context = list()
)
}
\arguments{
\item{id}{Module identifier.}

\item{registry}{Capability registry.}

\item{initial_graph}{Initial workflow graph.}

\item{context}{Host execution context.}
}
\description{
Capability workflow module server
}
