% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{default_capability_catalog}
\alias{default_capability_catalog}
\title{Create the exploratory default capability catalog}
\usage{
default_capability_catalog()
}
\description{
Create the exploratory default capability catalog
}
